LIB: Find APP starting address in the APP vector table
Checks that the stack pointer in the vector table points to
RAM (here generously considered to be up to 128k) thus between
0x20000000 and 0x2001ffff, to confirm that this is a valid
vector table.

Then looks up the reset handler in the vector table and jumps
to it.

This avoids the hardcoded address of the reset handler in the
branch table. Only the vector table location (0x800C000) of
the APP is now hardcoded in the LIB.

To not shift the entries in the branch table, a dummy function is
kept, which is also exported so it can be used to find the branch
table address in the symbol table.

If no valid APP (stack pointer really) is found, a message
is displayed.