Migrate vector table from assembly file to c file
There is no need to code it in assembly, and we should
rather try to maintain as much as possible in c.
The new stm32f10x_vector.c is almost the same as the IAR one.
I wonder why ST prefers to use so much assembly for gcc.

Incidentally, this finally makes the whole firmware work :)