GPIO_Config: Configure GPIO PA1 as analog input (Ain)
On Nano V2 both PA1 and PA0 are connected to Ain. So both should
be configured as analog input, otherwise unnecessary current would
flow through the pull-up resistor of PA1.

On Nano V1, PA1 is floating, so it should better be pulled up.
However, until we introduce some autodetection of this, we will just
favour the V2 hardware.