iar: Make a separate linker segment for the branch table
Create a segment following immediately the interrupt vector
table. As long as the vector table length is constant, we
know where this segment will know located. Put the branch
table here.

Update the APP ASM_Function.s hardcoded addresses to match
the new address of the branch table.

This is not the final solution but should work in most cases.